refactor(swingset): move kernel stats into own file Strongly type stats
