+++ packages/SwingSet/src/kernel/metrics.js
@@ -1,5 +1,7 @@
+// @ts-check
+
 // All the kernel metrics we are prepared for.
-export const KERNEL_STATS_SUM_METRICS = [
+export const KERNEL_STATS_SUM_METRICS = /** @type {const} */ ([
   {
     key: 'syscalls',
     name: 'swingset_syscall_total',
@@ -70,9 +72,9 @@ export const KERNEL_STATS_SUM_METRICS = [
     name: 'swingset_dispatch_notify_total',
     description: 'Total number of SwingSet vat promise notifications',
   },
-];
+]);
 
-export const KERNEL_STATS_UPDOWN_METRICS = [
+export const KERNEL_STATS_UPDOWN_METRICS = /** @type {const} */ ([
   {
     key: 'kernelObjects',
     name: 'swingset_kernel_objects',
@@ -131,10 +133,12 @@ export const KERNEL_STATS_UPDOWN_METRICS = [
     name: 'swingset_vats',
     description: 'Number of active vats',
   },
-];
+]);
 
-/** @type {ReadonlyArray<{key: string, name: string, consensus?: boolean, description: string, metricType: 'counter' | 'gauge'}>} */
-export const KERNEL_STATS_METRICS = [
-  ...KERNEL_STATS_SUM_METRICS.map(m => ({ ...m, metricType: 'counter' })),
-  ...KERNEL_STATS_UPDOWN_METRICS.map(m => ({ ...m, metricType: 'gauge' })),
-];
+const COUNTER = /** @type {const} */ ('counter');
+const GAUGE = /** @type {const} */ ('gauge');
+
+export const KERNEL_STATS_METRICS = harden([
+  ...KERNEL_STATS_SUM_METRICS.map(m => ({ ...m, metricType: COUNTER })),
+  ...KERNEL_STATS_UPDOWN_METRICS.map(m => ({ ...m, metricType: GAUGE })),
+]);
